+++ app/src/main/java/com/example/mobile_development_2_2/Map/Route/Route.kt
@@ -167,6 +167,17 @@ class Route(
         fun getSelectedPOI(): POI {
             return selectedItem
         }
+
+        fun getNextPOI(): POI? {
+            // Find the first not visited POI
+            for (poi in TestRoute().POIs) {
+                if (!poi.visited) {
+                    return poi
+                }
+            }
+
+            return null
+        }
     }
 
 }
