+++ app/src/main/java/com/example/mobile_development_2_2/data/GeofenceBroadcastReceiver.kt
@@ -1,6 +1,8 @@
 package com.example.mobile_development_2_2.data
 
+import android.app.PendingIntent
 import android.content.BroadcastReceiver
+import android.content.ContentValues
 import android.content.ContentValues.TAG
 import android.content.Context
 import android.content.Intent
@@ -10,15 +12,47 @@ import com.example.mobile_development_2_2.map.route.POI
 import com.example.mobile_development_2_2.gui.MainActivity
 import com.example.mobile_development_2_2.gui.fragments.MapFragment
 import com.example.mobile_development_2_2.map.route.RouteManager
-import com.google.android.gms.location.Geofence
-import com.google.android.gms.location.GeofenceStatusCodes
-import com.google.android.gms.location.GeofencingEvent
+import com.google.android.gms.location.*
 
 class GeofenceBroadcastReceiver : BroadcastReceiver() {
+
+    private lateinit var context: Context;
+
+    private var geofencingClient: GeofencingClient = LocationServices.getGeofencingClient(context)
+    private var geofenceHelper: GeofenceHelper = GeofenceHelper(context)
+
+    fun setGeofenceLocation(lat: Double, lng: Double, id : String  ) {
+        geofenceHelper.getPendingIntent()?.let { geofencingClient.removeGeofences(it) }
+        var geofence: Geofence? = geofenceHelper.getGeofence(id, lat, lng)
+
+        var geofencingRequest: GeofencingRequest? = geofence?.let {
+            geofenceHelper.geofencingRequest(
+                it
+            )
+        }
+        var pendingIntent: PendingIntent? = geofenceHelper.getPendingIntent()
+        if (geofencingRequest != null) {
+            if (pendingIntent != null) {
+                geofencingClient.addGeofences(geofencingRequest, pendingIntent)
+                    .addOnSuccessListener {
+                        Log.d(
+                            ContentValues.TAG,
+                            "Geofence added " + geofencingRequest.geofences[0].latitude + " " + geofencingRequest.geofences[0].longitude
+                        )
+                    }
+                    .addOnFailureListener { e ->
+                        Log.d(ContentValues.TAG, "onFailure: " + geofenceHelper.getErrorString(e))
+                    }
+            }
+        }
+    }
+
 
     private val TAG = "GeofenceBroadcastReceiver"
     // ...
     override fun onReceive(context: Context, intent: Intent) {
+        this.context = context
+
         var notificationHelper = NotificationHelper(context)
 
 
@@ -39,6 +73,12 @@ class GeofenceBroadcastReceiver : BroadcastReceiver() {
             for (geofence in geofenceList) {
                 Log.d(TAG, "onReceive: " + geofence.requestId + " triggered ")
                 notificationHelper.sendHighPriorityNotification("Geofence triggered",geofence.requestId, MainActivity::class.java)
+
+                // find next POI
+                val poi : POI? = RouteManager.getNextPOI()
+                if(poi != null) {
+                    setGeofenceLocation(poi.location.latitude, poi.location.longitude, poi.name)
+                }
 
                 //RouteManager.SelectPOI()
                 PopupHelper.SetState(true)
