+++ app/src/main/java/com/example/mobile_development_2_2/data/GeofenceManager.kt
@@ -1,0 +1,14 @@
+package com.example.mobile_development_2_2.data
+
+import android.app.PendingIntent
+import android.content.ContentValues
+import android.content.Context
+import android.util.Log
+import com.google.android.gms.location.Geofence
+import com.google.android.gms.location.GeofencingClient
+import com.google.android.gms.location.GeofencingRequest
+import com.google.android.gms.location.LocationServices
+
+class GeofenceManager(context : Context) {
+
+}
