+++ app/src/main/java/com/example/mobile_development_2_2/map/route/RouteManager.kt
@@ -53,5 +53,9 @@ class RouteManager {
         fun getSelectedPOI(): POI{
             return Route.getSelectedPOI()
         }
+
+        fun getNextPOI(): POI? {
+            return Route.getNextPOI()
+        }
     }
 }
