+++ app/src/main/java/com/example/mobile_development_2_2/ui/viewmodels/OSMViewModel.kt
@@ -33,9 +33,6 @@ import org.osmdroid.views.overlay.mylocation.IMyLocationProvider
 
 class OSMViewModel(getLocationProvider: GetLocationProvider, context : Context) : ViewModel() {
 
-
-    private var geofencingClient: GeofencingClient = LocationServices.getGeofencingClient(context)
-    private var geofenceHelper: GeofenceHelper = GeofenceHelper(context)
     var currentLocation : Location? = null
     private val lastLocations = getLocationProvider().shareIn(
         scope = viewModelScope,
@@ -101,42 +98,6 @@ class OSMViewModel(getLocationProvider: GetLocationProvider, context : Context) 
         coroutineScope = viewModelScope,
         locationFlow = lastLocations,
     )
-
-    fun invoke () {
-        for (it in pois) {
-            setGeofenceLocation(it.location.latitude,it.location.longitude, it.name)
-        }
-        //this method changes the location of the geofence,
-        //keep in mind there is always 1 active geofence which should be the next geofence in the route,
-        //"ïd" in this method should be the name of the geofence, this will be shown in the notification
-        setGeofenceLocation(51.5856, 4.7925, "geo")
-    }
-
-    fun setGeofenceLocation(lat: Double, lng: Double, id : String  ) {
-        geofenceHelper.getPendingIntent()?.let { geofencingClient.removeGeofences(it) }
-        var geofence: Geofence? = geofenceHelper.getGeofence(id, lat, lng)
-
-        var geofencingRequest: GeofencingRequest? = geofence?.let {
-            geofenceHelper.geofencingRequest(
-                it
-            )
-        }
-        var pendingIntent: PendingIntent? = geofenceHelper.getPendingIntent()
-        if (geofencingRequest != null) {
-            if (pendingIntent != null) {
-                geofencingClient.addGeofences(geofencingRequest, pendingIntent)
-                    .addOnSuccessListener {
-                        Log.d(
-                            ContentValues.TAG,
-                            "Geofence added " + geofencingRequest.geofences[0].latitude + " " + geofencingRequest.geofences[0].longitude
-                        )
-                    }
-                    .addOnFailureListener { e ->
-                        Log.d(ContentValues.TAG, "onFailure: " + geofenceHelper.getErrorString(e))
-                    }
-            }
-        }
-    }
 
     class Provider(
         private val coroutineScope: CoroutineScope,
